Add FakeT monad transformer
This is a use-case that came up in my work, where we were generating
lots of 'Fake' values. The trouble with running 'Fake' is that parsing
YAML and building the cache are very expensive, but all the effort
involved is thrown away after the 'Fake' is run - so it makes sense to
generate as much data as possible all in the same 'Fake'. But if you're
using a monad stack, you would have to put 'Fake' at the bottom - when
instead it makes more sense as a transformer, which can share a single
YAML parser and cache between many 'Fake' runs.

This adds the 'FakeT' monad transformer, and makes 'Fake' a synonym for
@FakeT IO@, preserving as much backwards-compatibility as possible with
functions and pattern synonyms.