Add MonadFake class for lifting Fakes
Following on from the 'FakeT' transformer, it makes sense to be able to
"lift" the default 'Fake' values into the monad stack, in order to allow
them to actually share the transformer cache and generator. This is done
by the 'MonadFake' class, which comes with an instance for 'FakeT', as
well as some of the other transformers defined in the `transformers`
package.